backend/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Muammar Al-Shedivat <muammar.alshedivat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

django = "^3.2.2"
django-admin-autocomplete-filter = "^0.7.1"
django-admin-list-filter-dropdown = "^1.0.3"
django-admin-rangefilter = "^0.7.0"
django-admin-sortable2 = "^1.0"
django-appconf = "^1.0.5"
django-autocomplete-light = "^3.9.4"
django-celery-beat = "^2.2.1"
django-colorfield = "^0.4.2"
django-cors-headers = "^3.13.0"
django-environ = "^0.4.5"
django-filter = "^21.1"
django-imagekit = "^4.0.2"
django-inline-actions = "^2.4.0"
django-model-utils = "^4.1.1"
django-multiselectfield = "^0.1.12"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.0.0"}
django-timezone-field = "^4.1.2"
django-safedelete = "^1.0.0"
django-simple-history = "^3.0.0"
djangorestframework = "^3.12.2"
drf-extra-fields = "^3.1.1"
drf-friendly-errors = { git = "https://github.com/citixensas/drf-friendly-errors.git" }
drf-spectacular = "^0.23.1"
fcm-django = "^0.3.7"
celery = "^5.2.1"
channels = "^3.0.4"
Collectfast = "^2.2.0"
Babel = "^2.9.1"
boto3 = "^1.18.1"
filetype = "^1.0.10"
firebase-admin = "^5.0.0"
freezegun = "^1.2.1"
openpyxl = "^3.0.7"
openrouteservice = "^2.3.3"
Pillow = "^8.0.1"
pydantic = "^1.9.1"
python-dateutil = "^2.8.2"
requests = "^2.25.1"
tenacity = "^7.0.0"
XlsxWriter = "^1.3.7"

# Celery
scikit-learn = "^1.1.2"
folium = "^0.12.1"

# Remove this line once the problem with 22.1.0 is fixed on OSX.
pyzmq = "<22.1.0"

# Production extras
psycopg2 = { version = "^2.8.6", optional = true }
gunicorn = { version = "^20.0.4", optional = true, extras = ["gevent"] }
sentry-sdk = { version = "^0.19.5", optional = true }
channels_redis = { version = "^3.3.1", optional = true }
qrcode = "^7.3.1"
djangorestframework-jwt = "^1.11.0"
djangorestframework-simplejwt = "^5.2.0"
react = "^4.3.0"
pywhatkit = "^5.4"

[tool.poetry.extras]
prod = ["psycopg2", "gunicorn", "sentry-sdk", "channels_redis"]
postgres = ["psycopg2"]

[tool.poetry.dev-dependencies]
# TODO: Add more linting dependencies
# Lint
pre-commit = "^2.9.3"
flake8 = "^3.8.4"
autoflake = "^1.4"
black = "^20.8b1"
isort = "^5.6.3"
toml = "^0.10.0"

# Testing
unittest-xml-reporting = "^2.5.2"
# Pickling tracebacks
tblib = "^1.7.0"

# Django
factory-boy = "^2.12.0"
parameterized = "^0.7.0"

# Performance & debugging
django-silk = "^4.1.0"
locust = "^1.2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
